Move formatting function to util file
