+++ src/app/view-debts/DebtsClient.tsx
@@ -1,8 +1,10 @@
 'use client'
 
 import React, { useEffect, useState } from 'react'
-import { getAllDebts, DebtEntry } from '@/lib/pint-bot-api/debts'
+import { getAllDebts } from '@/lib/pint-bot-api/debts'
 import { useDiscordUsers } from '@/hooks/useDiscordUsers'
+import { DebtEntry } from '@/lib/pint-bot-api/utils'
+import { formatMixed } from '@/lib/utils/formatting'
 
 export default function DebtsClient() {
   const [debts, setDebts] = useState<DebtEntry[]>([])
@@ -38,36 +40,6 @@ export default function DebtsClient() {
   }
   if (errorUsers) {
     return <p className="p-8 text-red-500">Error loading user names: {errorUsers}</p>
-  }
-
-  /**
- * Turn a string like "91/6" into "15 1/6",
- * "7/3" into "2 1/3", "4/2" into "2", "3/4" stays "3/4", and
- * plain numbers pass through unchanged.
- */
-  function formatMixed(fraction: string): string {
-    // if it’s not a fraction, just return it
-    if (!fraction.includes('/')) return fraction;
-
-    const [numStr, denStr] = fraction.split('/');
-    const num = parseInt(numStr, 10);
-    const den = parseInt(denStr, 10);
-
-    if (isNaN(num) || isNaN(den) || den === 0) return fraction;
-
-    const whole = Math.floor(num / den);
-    const rem = num % den;
-
-    if (rem === 0) {
-      // exact division
-      return String(whole);
-    }
-    if (whole === 0) {
-      // proper fraction
-      return `${rem}/${den}`;
-    }
-    // mixed number
-    return `${whole} ${rem}/${den}`;
   }
 
   return (
